[accumulator] accumulator flush unit test

+++ core/accumulator/src/node_index.rs
@@ -265,7 +265,7 @@ fn isolate_rightmost_zero_bit(v: u64) -> u64 {
 /// Turn off n right most bits
 fn turn_off_right_most_n_bits(v: u64, n: u32) -> u64 {
     precondition!(n < 64);
-    (v >> n) << n
+    (v >> n as u64) << n as u64
 }
 
 #[derive(Debug)]
